pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rf-detr-sagemaker"
version = "1.0.0"
description = "Real-Time Object Detection with RF-DETR and Amazon SageMaker"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "inference>=0.62.5",
    "numpy",
    "pillow",
    "rfdetr>=1.3.0",
    "sagemaker>=2",
    "supervision",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["code*", "utils*"]
exclude = ["tests*", "locust_scripts*", "local_inference*", "docker*", "previews*"]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
    "pytest",
    "pytest-cov",
    "pre-commit",
]
markdown = [
    "markdownlint-cli2",
]
locust = [
    "boto3",
    "botocore",
    "locust",
    "Pillow",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["rfdetr"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=code",
    "--cov=local_inference",
    "--cov-report=term-missing",
    "--cov-report=html",
]